shipyard_blueprints/bigquery/pyproject.toml
[tool.poetry]
name = "shipyard-bigquery"
version = "0.1.1a8"
description = "A local client for connecting to and working with Google BigQuery"
authors = ["wrp801 <wespoulsen@gmail.com>"]
license = "Apache 2.0"
readme = "README.md"
packages = [{include = "shipyard_bigquery"}]

[tool.poetry.dependencies]
python = ">=3.9, <3.10"
pandas = "^2.0"
google-cloud-bigquery = "^3.17.2"
db-dtypes = "1.1.1"
shipyard-utils = "^0.1.4"
shipyard-templates = "^0.6.1"
shipyard-bp-utils = "^1.0.1"


[tool.poetry.group.dev.dependencies]
shipyard-templates = {path = "../../shipyard-templates", develop = true}
python-dotenv = "^1.0.0"
pytest = "^7.4.0"
black = "^24.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
